+++ demo/index.js
@@ -19,18 +19,30 @@ var Editor = React.createClass({
 		this.setState({ value:value });
 	},
 
+	onReset: function() {
+		this.setState({ enabled: !this.state.enabled });
+	},
+
 	render: function() {
+		var quillComponent;
+		if (this.state.enabled) {
+			quillComponent = ReactQuill({
+				theme: 'snow',
+				value: this.state.value,
+				onChange: this.onEditorChange
+			});
+		}
 		return (
 			React.DOM.div({},
+				React.DOM.button({
+					onClick: this.onReset },
+					'Reset'
+				),
 				React.DOM.textarea({
 					value: this.state.value,
-					onChange: this.onTextareaChange
+					readOnly: true
 				}),
-				ReactQuill({
-					theme: 'snow',
-					defaultValue: this.state.value,
-					onChange: this.onEditorChange
-				})
+				quillComponent
 			)
 		);
 	}