+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 - Switched from `quilljs` package to `quill`.
+- Using the new `destroy()` from Quill.
 
 v0.0.2
 ------